Some modifications to ease data population
Exposed two fields to make it possible to define the order of multiple
series in the TimeTableXYDataset before adding data. This is useful if
you do know the order in which you want your series to appear but don't
know the order in which individual data points will be inserted into the
data set which may result in unpredicted order of series.